[boo] support input permutation in layer norm
Add an option to support input permutations in the signature and
generate sample arguments accordingly.

Refactor the Permutation class from conv into a generic utils.
